Update: Update-Event v3

+++ src/main/java/com/group02/openevent/repository/IFestivalEventRepo.java
@@ -3,5 +3,5 @@ package com.group02.openevent.repository;
 import com.group02.openevent.model.event.FestivalEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IFestivalEventRepo extends JpaRepository<FestivalEvent, Integer> {
+public interface IFestivalEventRepo extends JpaRepository<FestivalEvent, Long> {
 }
